# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural Portfolio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
        - name: category_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
  /projects/{id}/update:
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/publish:
    patch:
      operationId: toggleProjectPublish
      tags: [projects]
      summary: Toggle project published status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/delete:
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /machinery:
    get:
      operationId: listMachinery
      tags: [machinery]
      summary: List all machinery
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of machinery
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MachineryItem"
    post:
      operationId: createMachinery
      tags: [machinery]
      summary: Create machinery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
  /machinery/{slug}:
    get:
      operationId: getMachinery
      tags: [machinery]
      summary: Get machinery by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Machinery detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
        "404":
          description: Not found
  /machinery/{id}/update:
    put:
      operationId: updateMachinery
      tags: [machinery]
      summary: Update machinery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
  /machinery/{id}/publish:
    patch:
      operationId: toggleMachineryPublish
      tags: [machinery]
      summary: Toggle machinery published status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineryItem"
  /machinery/{id}/delete:
    delete:
      operationId: deleteMachinery
      tags: [machinery]
      summary: Delete machinery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ProjectImage:
      type: object
      required: [id, projectId, imageUrl, isHero, sortOrder]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        imageUrl:
          type: string
        isHero:
          type: boolean
        sortOrder:
          type: integer
...
    ProjectSummary:
      type: object
      required: [id, title, slug, status, published]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        location:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        status:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        heroImage:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
    Project:
      type: object
      required: [id, title, slug, status, published]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        location:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        status:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
        longDescription:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        heroImage:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProjectImage"
    ProjectInput:
      type: object
      required: [title, slug, status]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        location:
          type: string
        client:
          type: string
        sector:
          type: string
        size:
          type: string
        scope:
          type: string
        status:
          type: string
        longDescription:
          type: string
        categoryId:
          type: ["integer", "null"]
        serviceId:
          type: ["integer", "null"]
        year:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
    PublishToggle:
      type: object
      required: [published]
      properties:
        published:
          type: boolean
    MachineryItem:
      type: object
      required: [id, name, slug, published]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        longDescription:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        galleryImages:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        condition:
          type: ["string", "null"]
        published:
          type: boolean
        featured:
          type: boolean
    MachineryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        category:
          type: string
        description:
          type: string
        longDescription:
          type: string
        imageUrl:
          type: string
        galleryImages:
          type: string
        year:
          type: string
        condition:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
...