# lib/api-zod/src/generated/types/machineryInput.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface MachineryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  category?: string;
  description?: string;
  longDescription?: string;
  imageUrl?: string;
  galleryImages?: string;
  year?: string;
  condition?: string;
  published?: boolean;
  featured?: boolean;
}